Skripte/requirements.txt
streamlit
pandas
opencv-python-headless
pytesseract
Pillow
jiwer
scikit-image
transformers
# PyTorch separat installieren (nicht hier!)
